refactor(dashboard): split the stat-card and topology-graph pieces out (#1040)
Relocations only, no behaviour change. Both source files sit exactly at their
file-budget ceiling, and #1040 needs to add a node-location row to the Tari card
and a remote_tari input to the topology builder. The gate's own answer to that is
to make the file smaller rather than to raise the bar, so this takes the pieces
that carry no dependency back on their old home:

- static/statcards.mjs <- StatCard, MoreStats and TariStatus. components.mjs is a
  hub that already imports chart.mjs, topology.mjs and configview.mjs; these three
  are leaves, so the import runs one way and closes no cycle. 1400 -> 1350 lines.
- service/topology_graph.py <- the zone constants and TOPOLOGY_NODES, which are
  data rather than logic. The zones are used by nothing else, so the block moves
  whole. egress.py re-exports TOPOLOGY_NODES, so every existing importer —
  including tests/service/test_egress.py — keeps working untouched. 435 -> 413.

topology_graph also gains topology_nodes(), unused as of this commit and wired up
in the next one.

Verified unchanged: node --test on the frontend tier (452 pass, same as before),
tests/service/test_egress.py, make lint-js and make lint-py.
